+++ R/utils.R
@@ -321,7 +321,7 @@ h_get_na_action <- function(na_action) {
     return(na_action)
   }
   if (is.character(na_action) && length(na_action) == 1L) {
-    assert_subset(na_action, c("na.omit", "na.exclude", "na.fail", "na.pass", "na.exclude"))
+    assert_subset(na_action, c("na.omit", "na.exclude", "na.fail", "na.pass", "na.contiguous"))
     return(get(na_action, mode = "function", pos = "package:stats"))
   }
 }